Remove initial commit creation
Somehow, after the initial commit, extra file changes are being generated.
This leaves the repository in a confusing state after running `bretels`.

Apart from this, I don't think it should be bretels' responsibility to do the
initial commit for us. Feels better to leave this to the user running bretels.
Configuring sane defaults for .gitignore and doing `git init` is fine.